make a util for page selection
